# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "CanvasMaterialType": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "emoji": {
            "default": "",
            "title": "Emoji",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "structure": {
            "default": "",
            "title": "Structure",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "category"
        ],
        "title": "CanvasMaterialType",
        "type": "object"
      },
      "CanvasMaterialTypesPayload": {
        "properties": {
          "material_types": {
            "items": {
              "$ref": "#/components/schemas/CanvasMaterialType"
            },
            "title": "Material Types",
            "type": "array"
          }
        },
        "required": [
          "material_types"
        ],
        "title": "CanvasMaterialTypesPayload",
        "type": "object"
      },
...
      "ContextActionsConfig": {
        "properties": {
          "curate_prompt": {
            "title": "Curate Prompt",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "greetings": {
            "$ref": "#/components/schemas/ContextGreetings"
          },
          "pills": {
            "$ref": "#/components/schemas/ContextPills"
          },
          "report_url": {
            "title": "Report Url",
            "type": "string"
          }
        },
        "required": [
          "report_url",
          "greetings",
          "pills",
          "curate_prompt"
        ],
        "title": "ContextActionsConfig",
        "type": "object"
      },
      "ContextGreetings": {
        "properties": {
          "collection": {
            "title": "Collection",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "collection",
          "content",
          "topic"
        ],
        "title": "ContextGreetings",
        "type": "object"
      },
      "ContextPill": {
        "properties": {
          "action": {
            "default": "",
            "title": "Action",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label",
          "kind"
        ],
        "title": "ContextPill",
        "type": "object"
      },
      "ContextPills": {
        "properties": {
          "collection": {
            "items": {
              "$ref": "#/components/schemas/ContextPill"
            },
            "title": "Collection",
            "type": "array"
          },
          "content": {
            "items": {
              "$ref": "#/components/schemas/ContextPill"
            },
            "title": "Content",
            "type": "array"
          },
          "topic": {
            "items": {
              "$ref": "#/components/schemas/ContextPill"
            },
            "title": "Topic",
            "type": "array"
          }
        },
        "required": [
          "collection",
          "content",
          "topic"
        ],
        "title": "ContextPills",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PrivacyConfig": {
        "properties": {
          "memory": {
            "default": true,
            "title": "Memory",
            "type": "boolean"
          },
          "messages": {
            "default": true,
            "title": "Messages",
            "type": "boolean"
          },
          "quality": {
            "default": true,
            "title": "Quality",
            "type": "boolean"
          },
          "safety": {
            "default": true,
            "title": "Safety",
            "type": "boolean"
          }
        },
        "title": "PrivacyConfig",
        "type": "object"
      },
...
      "ToneModifier": {
        "properties": {
          "card_text_mode": {
            "default": "minimal",
            "title": "Card Text Mode",
            "type": "string"
          },
          "formality": {
            "default": "wie_user",
            "title": "Formality",
            "type": "string"
          },
          "length_bias": {
            "default": 0.0,
            "title": "Length Bias",
            "type": "number"
          },
          "override": {
            "default": false,
            "title": "Override",
            "type": "boolean"
          },
          "tone": {
            "default": "locker",
            "title": "Tone",
            "type": "string"
          }
        },
        "title": "ToneModifier",
        "type": "object"
      },
      "ToneModifiersPayload": {
        "properties": {
          "default_modifier": {
            "$ref": "#/components/schemas/ToneModifier"
          },
          "modifiers": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ToneModifier"
            },
            "title": "Modifiers",
            "type": "object"
          }
        },
        "required": [
          "modifiers",
          "default_modifier"
        ],
        "title": "ToneModifiersPayload",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WelcomeConfig": {
        "properties": {
          "greeting": {
            "title": "Greeting",
            "type": "string"
          },
          "quick_replies": {
            "items": {
              "type": "string"
            },
            "title": "Quick Replies",
            "type": "array"
          },
          "tour_reply": {
            "default": "",
            "title": "Tour Reply",
            "type": "string"
          }
        },
        "required": [
          "greeting",
          "quick_replies"
        ],
        "title": "WelcomeConfig",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "StudioCookie": {
        "in": "cookie",
        "name": "boerdi_studio_auth",
        "type": "apiKey"
      },
      "StudioKey": {
        "in": "header",
        "name": "X-Studio-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "boerdi-chat backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/config/canvas/material-types": {
      "get": {
        "operationId": "get_canvas_material_types_api_config_canvas_material_types_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CanvasMaterialTypesPayload"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Get Canvas Material Types",
        "tags": [
          "config"
        ]
      },
      "put": {
        "operationId": "update_canvas_material_types_api_config_canvas_material_types_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CanvasMaterialTypesPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CanvasMaterialTypesPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Update Canvas Material Types",
        "tags": [
          "config"
        ]
      }
    },
    "/api/config/context-actions": {
      "get": {
        "operationId": "get_context_actions_config_api_config_context_actions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextActionsConfig"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Get Context Actions Config",
        "tags": [
          "config"
        ]
      },
      "put": {
        "operationId": "update_context_actions_config_api_config_context_actions_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextActionsConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextActionsConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Update Context Actions Config",
        "tags": [
          "config"
        ]
      }
    },
...
    "/api/config/privacy": {
      "get": {
        "operationId": "get_privacy_config_api_config_privacy_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyConfig"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Get Privacy Config",
        "tags": [
          "config"
        ]
      },
      "put": {
        "operationId": "update_privacy_config_api_config_privacy_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrivacyConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Update Privacy Config",
        "tags": [
          "config"
        ]
      }
    },
...
    "/api/config/tone-modifiers": {
      "get": {
        "operationId": "get_tone_modifiers_api_config_tone_modifiers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToneModifiersPayload"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Get Tone Modifiers",
        "tags": [
          "config"
        ]
      },
      "put": {
        "description": "Persona modifiers go into each persona area's frontmatter; the\ndefault stays in 01-base/tone-modifiers (fallback for unknown personas).",
        "operationId": "update_tone_modifiers_api_config_tone_modifiers_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToneModifiersPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToneModifiersPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Update Tone Modifiers",
        "tags": [
          "config"
        ]
      }
    },
    "/api/config/welcome": {
      "get": {
        "operationId": "get_welcome_config_api_config_welcome_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WelcomeConfig"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Get Welcome Config",
        "tags": [
          "config"
        ]
      },
      "put": {
        "operationId": "update_welcome_config_api_config_welcome_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WelcomeConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WelcomeConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Update Welcome Config",
        "tags": [
          "config"
        ]
      }
    },
...
    "/api/rag/area/{area}": {
      "delete": {
        "description": "Delete all chunks in a knowledge area, and its document rows.\n\nALT swept the sqlite-vec index here too, because embeddings lived in a\nseparate table whose rows could outlive their chunk. pgvector keeps the\nembedding as a column on ``rag_chunks``, so the vector dies with the chunk\nand there is nothing left to sweep (see services/rag/admin.py).",
        "operationId": "delete_rag_area_api_rag_area__area__delete",
        "parameters": [
          {
            "in": "path",
            "name": "area",
            "required": true,
            "schema": {
              "title": "Area",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Delete Rag Area",
        "tags": [
          "rag"
        ]
      },
      "get": {
        "description": "List documents in a knowledge area.\n\nDocuments are grouped by the compound key ``(title, source)`` so that\ne.g. two uploads with the same filename from different folders, or two\nmanual entries with the same title, remain distinguishable.",
        "operationId": "get_rag_area_api_rag_area__area__get",
        "parameters": [
          {
            "in": "path",
            "name": "area",
            "required": true,
            "schema": {
              "title": "Area",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Get Rag Area",
        "tags": [
          "rag"
        ]
      }
    },
...
  }
}